# shared/openapi.yaml
components:
  schemas:
    CancelRunResponse:
      additionalProperties: false
      description: Body of the 200 response from ``POST /api/runs/{id}/cancel``.
      properties:
        run_id:
          title: Run Id
          type: string
        status:
          description: The new status after the cancel flag is set.
          title: Status
          type: string
      required:
      - run_id
      - status
      title: CancelRunResponse
      type: object
    ComparisonReport:
      additionalProperties: false
      description: "Top-level Comparison_Report returned by :func:`compare`.\n\n``metric_diffs``\
        \ and ``performance_diffs`` are lists (not dicts)\nto match the design's \xA7\
        Data Models / Comparison_Report table and\nbecause list ordering is stable\
        \ across re-renderings in the UI\n(Property 41). Ordering is deterministic:\
        \ the lists are sorted\nlexicographically by key tuple so two independent\
        \ invocations on\nequal inputs produce equal reports."
      properties:
        metric_diffs:
          description: Per-``(model, metric)`` diffs over the intersection.
          items:
            $ref: '#/components/schemas/MetricDiff'
          title: Metric Diffs
          type: array
        performance_diffs:
          description: Per-model performance diffs over the intersection.
          items:
            $ref: '#/components/schemas/PerformanceDiff'
          title: Performance Diffs
          type: array
        run_a:
          description: '``RunReport.run_id`` of the base Run.'
          title: Run A
          type: string
        run_b:
          description: '``RunReport.run_id`` of the comparison Run.'
          title: Run B
          type: string
      required:
      - run_a
      - run_b
      - metric_diffs
      - performance_diffs
      title: ComparisonReport
      type: object
...
    EvaluationSuite:
      additionalProperties: false
      description: 'A named, versioned collection of :class:`TestCase` objects.


        A suite is the primary on-disk artifact authored by users (YAML or JSON;

        see ``suites/loader.py``) and the unit of filtering at Run time

        (Requirement 3.6). Suites must have a non-empty ``name``, a non-empty

        ``test_cases`` list, and every ``test_cases[i].id`` must be unique

        within the suite (Requirement 3.3).'
      properties:
        defaults:
          $ref: '#/components/schemas/GenerationDefaults'
          description: Default generation parameters for every TestCase in this suite.
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional free-form description shown in the UI.
          title: Description
        name:
          description: Human-readable suite name, used by ``RunConfig.suites``.
          title: Name
          type: string
        test_cases:
          description: Ordered, non-empty list of test cases with unique ids.
          items:
            $ref: '#/components/schemas/TestCase'
          title: Test Cases
          type: array
        version:
          default: '1.0'
          description: Authoring version string; informational only in v1.
          title: Version
          type: string
      required:
      - name
      - test_cases
      title: EvaluationSuite
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MetricDiff:
      additionalProperties: false
      description: "One row of :class:`ComparisonReport.metric_diffs`.\n\n``diff``\
        \ is ``mean_b - mean_a``. Positive means Run B scored\nhigher than Run A on\
        \ the given metric; negative means the\nopposite. The field is a ``float``\
        \ (not ``float | None``)\nbecause the enclosing keyset is the *intersection*\
        \ \u2014 a\n``MetricDiff`` is only emitted when both reports have a value."
      properties:
        diff:
          description: '``mean_b - mean_a``.'
          title: Diff
          type: number
        mean_a:
          description: Mean score in Run A.
          title: Mean A
          type: number
        mean_b:
          description: Mean score in Run B.
          title: Mean B
          type: number
        metric:
          description: Metric name.
          title: Metric
          type: string
        model:
          description: Model tag the metric was scored for.
          title: Model
          type: string
      required:
      - model
      - metric
      - mean_a
      - mean_b
      - diff
      title: MetricDiff
      type: object
...
    PerformanceDiff:
      additionalProperties: false
      description: 'One row of :class:`ComparisonReport.performance_diffs`.


        Populated per model present in both reports. ``*_tps_*`` fields

        are nullable because the underlying per-model mean is nullable

        (Requirement 6.5). ``tps_diff`` is ``None`` when either side is

        ``None``; ``total_ms_diff`` is always a concrete ``float``

        because ``mean_total_ms`` is non-nullable on

        :class:`ModelAggregate`.'
      properties:
        mean_tokens_per_second_a:
          anyOf:
          - type: number
          - type: 'null'
          description: Run A's mean tokens-per-second for this model.
          title: Mean Tokens Per Second A
        mean_tokens_per_second_b:
          anyOf:
          - type: number
          - type: 'null'
          description: Run B's mean tokens-per-second for this model.
          title: Mean Tokens Per Second B
        mean_total_ms_a:
          description: Run A's mean total response time for this model (ms).
          title: Mean Total Ms A
          type: number
        mean_total_ms_b:
          description: Run B's mean total response time for this model (ms).
          title: Mean Total Ms B
          type: number
        model:
          description: Model tag.
          title: Model
          type: string
        total_ms_diff:
          description: '``mean_total_ms_b - mean_total_ms_a``.'
          title: Total Ms Diff
          type: number
        tps_diff:
          anyOf:
          - type: number
          - type: 'null'
          description: '``mean_tokens_per_second_b - mean_tokens_per_second_a``; ``None``
            when either side is ``None``.'
          title: Tps Diff
      required:
      - model
      - mean_tokens_per_second_a
      - mean_tokens_per_second_b
      - mean_total_ms_a
      - mean_total_ms_b
      - tps_diff
      - total_ms_diff
      title: PerformanceDiff
      type: object
...
    RunConfig:
      additionalProperties: false
      description: 'Run-level configuration for a single Run.


        ``RunConfig`` is the body the Backend_API accepts for

        ``POST /api/runs`` and is also embedded in :class:`ConfigFile` as the

        ``run:`` section. A ``RunConfig`` is serialised verbatim into every

        ``RunReport`` so runs are reproducible from their report alone

        (Requirement 8.4).


        Validators enforce the numeric ranges specified in the Requirements

        document and reject empty ``models`` / ``suites`` lists early, before

        any Ollama_Server call is made.'
      properties:
        concurrency:
          default: 1
          description: Maximum number of in-flight Ollama_Server requests during this
            Run. Must be >= 1 (Requirement 5.5).
          title: Concurrency
          type: integer
        judge_model:
          anyOf:
          - type: string
          - type: 'null'
          description: Ollama model tag used by any configured ``llm-as-judge`` metric.
            ``None`` means no judge model is configured; metrics that require one
            will raise a validation error at Run start.
          title: Judge Model
        models:
          description: Ollama model tags to evaluate in this Run (e.g. 'llama3:8b').
            Must contain at least one model (Requirement 2.2).
          items:
            type: string
          title: Models
          type: array
        ollama_timeout_s:
          default: 120.0
          description: Per-request timeout in seconds applied to every Ollama_Server
            call made during this Run (Requirement 1.4). Must be > 0.
          title: Ollama Timeout S
          type: number
        pull_missing_models:
          default: false
          description: If true, instruct the Ollama_Server to pull any requested Model
            that is not already available (Requirement 2.4).
          title: Pull Missing Models
          type: boolean
        repetitions:
          default: 1
          description: Number of times each Test_Case is executed per Model. Must
            be >= 1 (Requirement 5.2).
          title: Repetitions
          type: integer
        retry_max_attempts:
          default: 2
          description: Number of *additional* retry attempts per Test_Case execution
            on retryable network errors (Requirement 11.1). Must be >= 0. A value
            of 0 disables retries entirely.
          title: Retry Max Attempts
          type: integer
        suites:
          description: Names of Evaluation_Suites to run. Must contain at least one
            name (Requirement 3.1).
          items:
            type: string
          title: Suites
          type: array
        tag_filter:
          description: Optional Test_Case tag filter (Requirement 3.6). When empty,
            every Test_Case in the selected suites is included. When non-empty, only
            Test_Cases whose ``tags`` intersect this list are selected.
          items:
            type: string
          title: Tag Filter
          type: array
      required:
      - models
      - suites
      title: RunConfig
      type: object
    RunReport:
      additionalProperties: false
      description: "Top-level Run_Report persisted to disk and returned by the REST\
        \ API.\n\nA ``RunReport`` is the single source of truth for a completed Run\n\
        (Requirement 8.4): it embeds the full submitted :class:`ConfigFile`,\nthe\
        \ identifying metadata for every evaluated model, every\nper-execution result,\
        \ the per-model aggregates, and an error\nsummary. Property 16 (Run_Report\
        \ completeness) asserts this\ninvariant; Property 18 (round-trip) asserts\
        \ that\n``RunReport.model_validate_json(r.model_dump_json()) == r`` for\n\
        every valid ``r``.\n\n``status`` is a five-value literal covering the full\
        \ state machine:\n\n* ``pending`` \u2014 the Run has been created but dispatch\
        \ has not started.\n* ``running`` \u2014 at least one execution is in flight\
        \ or queued.\n* ``completed`` \u2014 every planned execution has produced\
        \ a result.\n* ``aborted`` \u2014 the Run was stopped by a cancel request\
        \ or signal\n  before all planned executions completed (Requirement 11.4).\n\
        * ``failed`` \u2014 the Run could not start (e.g. missing models in\n  preflight,\
        \ Property 5) and no per-execution results exist.\n\nThe ``ended_at >= started_at``\
        \ invariant is enforced in a model\nvalidator rather than field-by-field because\
        \ it is a cross-field\nconstraint (Property 16)."
      properties:
        aggregates:
          description: One ``ModelAggregate`` per evaluated model (Requirement 8.2).
          items:
            $ref: '#/components/schemas/ModelAggregate'
          title: Aggregates
          type: array
        backend_version:
          description: Version string of the Backend that produced this report.
          title: Backend Version
          type: string
        config:
          $ref: '#/components/schemas/ConfigFile'
          description: Full submitted ``ConfigFile`` (Requirement 8.4).
        ended_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: UTC timestamp when the Run reached a terminal state. ``None``
            while the Run is still ``pending`` or ``running``.
          title: Ended At
        error_summary:
          description: One entry per result with ``status`` in ``{'error', 'timeout'}``
            (Requirement 11.3).
          items:
            $ref: '#/components/schemas/ErrorSummaryEntry'
          title: Error Summary
          type: array
        models:
          description: One ``ModelInfo`` per evaluated model (Requirement 2.5).
          items:
            $ref: '#/components/schemas/ModelInfo'
          title: Models
          type: array
        ollama_version:
          anyOf:
          - type: string
          - type: 'null'
          description: Ollama_Server version string at Run start. ``None`` when the
            Backend could not reach the server before failing.
          title: Ollama Version
        results:
          description: One ``TestCaseResult`` per ``(model, test_case, repetition)``.
          items:
            $ref: '#/components/schemas/TestCaseResult'
          title: Results
          type: array
        run_id:
          description: Unique identifier for this Run (Property 25). Must be non-empty.
          title: Run Id
          type: string
        started_at:
          description: UTC timestamp when the Run started.
          format: date-time
          title: Started At
          type: string
        status:
          description: Terminal or in-flight state of the Run (see class docstring).
          enum:
          - pending
          - running
          - completed
          - aborted
          - failed
          title: Status
          type: string
      required:
      - run_id
      - backend_version
      - ollama_version
      - started_at
      - ended_at
      - status
      - config
      - models
      - results
      - aggregates
      - error_summary
      title: RunReport
      type: object
    SubmitRunResponse:
      additionalProperties: false
      description: Body of the 201 response from ``POST /api/runs`` (Requirement 13.3).
      properties:
        run_id:
          description: Identifier of the newly queued Run.
          title: Run Id
          type: string
        status:
          description: Initial status; always ``pending``.
          title: Status
          type: string
      required:
      - run_id
      - status
      title: SubmitRunResponse
      type: object
...
info:
  description: Local evaluation harness for Ollama-hosted LLMs.
  title: Ollama Model Evaluator
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/compare:
    get:
      description: Return the comparison between two persisted Run_Reports.
      operationId: compare_runs_api_compare_get
      parameters:
      - description: Run id of the base run.
        in: query
        name: a
        required: true
        schema:
          description: Run id of the base run.
          title: A
          type: string
      - description: Run id of the comparison run.
        in: query
        name: b
        required: true
        schema:
          description: Run id of the comparison run.
          title: B
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonReport'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Compare Runs
...
  /api/runs:
    get:
      description: Return stored runs matching the supplied filters.
      operationId: list_runs_api_runs_get
      parameters:
      - in: query
        name: model
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
      - in: query
        name: suite
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Suite
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - in: query
        name: since
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Since
      - in: query
        name: until
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Until
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response List Runs Api Runs Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Runs
    post:
      description: Persist a new Run as ``pending`` and hand it to the supervisor.
      operationId: submit_run_api_runs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunConfig'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Submit Run
  /api/runs/{run_id}:
    delete:
      description: Delete a Run and its persisted artifacts (Requirement 12.5).
      operationId: delete_run_api_runs__run_id__delete
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Run
    get:
      description: Return the full :class:`RunReport` for ``run_id``.
      operationId: get_run_api_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunReport'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Run
  /api/runs/{run_id}/cancel:
    post:
      description: Request cooperative cancellation of a live Run.
      operationId: cancel_run_api_runs__run_id__cancel_post
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Cancel Run
  /api/runs/{run_id}/report.md:
    get:
      description: Return the Run_Report rendered as ``text/markdown``.
      operationId: get_run_markdown_api_runs__run_id__report_md_get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Run Markdown
...
  /api/suites/{name}:
    get:
      description: Return the full :class:`EvaluationSuite` matching ``name``.
      operationId: get_suite_api_suites__name__get
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationSuite'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Suite